# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MediaHub API
  description: |
    CQRS / Event Sourcing / Saga Pattern 学習用メディア管理プラットフォーム。

    すべての `/api/v1/*` エンドポイントは JWT 認証が必要です。
    開発時は `POST /auth/dev-token` で取得したトークンを `Authorization: Bearer <token>` ヘッダーに設定してください。

    ## アーキテクチャ
    - **Gateway (8080)**: 外部からの唯一のエントリポイント。JWT 認証後、内部サービスへプロキシ
    - **media-command (8081)**: メディアの書き込み操作（CQRS Command 側）
    - **media-query (8082)**: メディアの読み取り操作（CQRS Query 側、Read Model）
    - **album (8083)**: アルバム管理
    - **eventstore (8084)**: イベントの永続化と配信
    - **saga (8085)**: 分散トランザクション（Orchestration Saga）
    - **notification (8086)**: イベント駆動の通知
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: nao1215
    url: https://github.com/nao1215/micro
servers:
  - url: http://localhost:8080
    description: Gateway（開発環境）
...
paths:
...
  /api/v1/media:
    get:
      tags: [media]
      summary: メディア一覧取得
      description: 認証ユーザーのメディア一覧を Read Model から取得する。
      operationId: listMedia
      security:
        - bearerAuth: []
      responses:
        "200":
          description: メディア一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [media]
      summary: メディアアップロード
      description: |
        画像または動画ファイルをアップロードする（最大 50MB）。
        アップロード後、Saga によるサムネイル生成・アルバム追加が自動的に開始される。
      operationId: uploadMedia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 画像ファイル（image/*）または動画ファイル（video/*）
      responses:
        "201":
          description: アップロード成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"
        "400":
          description: 不正なリクエスト（ファイル未指定、サイズ超過、不正な Content-Type 等）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/media/{id}:
    get:
      tags: [media]
      summary: メディア詳細取得
      operationId: getMedia
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MediaId"
      responses:
        "200":
          description: メディア詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "404":
          description: メディアが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [media]
      summary: メディア削除
      description: メディアファイルとメタデータを削除し、MediaDeleted イベントを発行する。
      operationId: deleteMedia
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MediaId"
      responses:
        "200":
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: メディアが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/events:
    get:
      tags: [event]
      summary: イベントログ取得
      description: Event Store に記録された全イベントを取得する。
      operationId: listEvents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: イベント一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponse"

  # ============================================================
  # Event Store 内部 API（ポート 8084）
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `POST /auth/dev-token` で取得した JWT トークン。
        Claims: `{ "user_id": string, "email": string, "exp": number, "iat": number, "iss": "mediahub-gateway" }`
  parameters:
    MediaId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: メディア ID
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ
      example:
        error: "リクエストが不正です"
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      example:
        message: "操作が完了しました"
...
    MediaUploadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
          format: int64
        storage_path:
          type: string
    MediaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
          format: int64
        storage_path:
          type: string
        thumbnail_path:
          type: string
        width:
          type: integer
        height:
          type: integer
        duration_seconds:
          type: number
          format: double
        status:
          type: string
          enum: [uploaded, processed, failed, compensated]
        uploaded_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MediaListResponse:
      type: object
      properties:
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaResponse"
        count:
          type: integer
...
    EventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        aggregate_id:
          type: string
          format: uuid
        aggregate_type:
          type: string
        event_type:
          type: string
        data:
          type: object
          description: イベントデータ（JSON）
        version:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time